+++ tx-backend/src/main/java/org/eclipse/tractusx/traceability/common/properties/BpdmProperties.java
@@ -19,11 +19,9 @@
 package org.eclipse.tractusx.traceability.common.properties;
 
 import lombok.Getter;
-import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Configuration;
 
-@RequiredArgsConstructor
 @Getter
 @Configuration
 public class BpdmProperties {